backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0,<11.0
gunicorn>=20.1.0
mysqlclient>=2.1.0
python-dotenv>=1.0.0
coverage>=7.0,<8.0
